+++ cli.js
@@ -78,6 +78,12 @@ program
 		null,
 	)
 	.option(
+		'--uses <value>',
+		'The number of times the new URL can be used (set to 0 for unlimited uses)',
+		_.identity,
+		1,
+	)
+	.option(
 		'--params [values]',
 		'Stringified JSON object of params for the newly generated URL - e.g for "withdrawRequest" valid params could be {"minWithdrawable": 1000, "maxWithdrawable": 5000}',
 		_.identity,
@@ -147,11 +153,10 @@ program
 			if (options.store && options.store.backend === 'memory') {
 				throw new Error('This command does not work with `--store.backend` set to "memory"');
 			}
-			const { tag } = this;
+			let { tag, params, uses } = this;
 			if (!tag) {
 				throw new Error('--tag is required');
 			}
-			let { params } = this;
 			if (!params) {
 				params = {};
 			} else if (_.isString(params)) {
@@ -161,10 +166,14 @@ program
 					throw new Error('--params must be a valid JSON object');
 				}
 			}
+			uses = uses && parseInt(uses) || 1;
+			if (_.isNaN(uses)) {
+				throw new Error('--uses must be an integer');
+			}
 			options.listen = false
 			options.lightning = null;
 			const server = lnurl.createServer(options);
-			return server.generateNewUrl(tag, params).then(result => {
+			return server.generateNewUrl(tag, params, { uses }).then(result => {
 				process.stdout.write(JSON.stringify(result, null, 2));
 				process.exit();
 			}).catch(error => {