+++ server/stores/migrations/0002_urls_initialUses.js
@@ -1,0 +1,19 @@
+exports.up = function(knex) {
+	return knex.schema.hasColumn('urls', 'initialUses').then(exists => {
+		if (!exists) {
+			return knex.schema.table('urls', table => {
+				table.integer('initialUses').unsigned().defaultTo(1);
+			});
+		}
+	});
+};
+
+exports.down = function(knex) {
+	return knex.schema.hasColumn('urls', 'initialUses').then(exists => {
+		if (exists) {
+			return knex.schema.table('urls', table => {
+				table.dropColumn('initialUses');
+			});
+		}
+	});
+};